Feature/be/ns fix

See merge request s08-bigdata-recom-sub2/S08P22A503!182

Former-commit-id: bbf4b06e1c9b687920f7c6afa83b55ef5c59e5f6

+++ backend/src/main/java/com/a503/churros/repository/news/ReadRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ReadRepository extends JpaRepository<Read, Long> {
-    Optional<Read> findByUserIdxAndArticleIdx(long userIdx , long articleIdx);
+    Optional<Read> findByUserIdxAndArticleIdx(long userIdx , int articleIdx);
 }
